feat: moved the majority of the methods from numpy to asbtract
